to make the code more readable and understandable i split all the logic inside the chat.tsx file to a hooks file, utils file and a socket handler file
