+++ src/components/Chat.tsx
@@ -1,6 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { io, Socket } from 'socket.io-client';
+import React from 'react';
 import { 
   AppBar, Toolbar, Typography, Box, IconButton, Dialog, DialogTitle, 
   DialogContent, DialogActions, Snackbar, Avatar, TextField, Button, 
@@ -10,344 +8,48 @@ import {
   Settings as SettingsIcon, 
   UploadFile as UploadFileIcon
 } from '@mui/icons-material';
-import { API } from '../services/api';
-import { API_URL } from '../variables/Variables';
-import { Message, Group } from '../types/Types';
 import { ChatContainer } from '../styles/StyledComponents';
 import GroupList from './GroupList';
 import ChatRoom from './ChatRoom';
 import MemberList from './MemberList';
-
+import { useChatLogic } from '../hooks/useChatLogic';
 const Chat: React.FC = () => {
-  const [socket, setSocket] = useState<Socket | null>(null);
-  const [userId, setUserId] = useState('');
-  const [selectedRoom, setSelectedRoom] = useState<string | null>(null);
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [inputMessage, setInputMessage] = useState('');
-  const [groups, setGroups] = useState<Group[]>([]);
-  const [isProfilePictureDialogOpen, setIsProfilePictureDialogOpen] = useState(false);
-  const [isGroupPictureDialogOpen, setIsGroupPictureDialogOpen] = useState(false);
-  const [newProfilePictureUrl, setNewProfilePictureUrl] = useState('');
-  const [newGroupPictureUrl, setNewGroupPictureUrl] = useState('');
-  const [profilePicture, setProfilePicture] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const messageListRef = useRef<HTMLUListElement>(null);
-  const navigate = useNavigate();
-  const [socketError, setSocketError] = useState<string | null>(null);
-  const [roomMembers, setRoomMembers] = useState<string[]>([]);
-  const [userNames, setUserNames] = useState<{[key: string]: string}>({});
-  const [currentUserName, setCurrentUserName] = useState<string>('');
+  const {
+    userId,
+    selectedRoom,
+    messages,
+    inputMessage,
+    setInputMessage,
+    groups,
+    setGroups,
+    isProfilePictureDialogOpen,
+    setIsProfilePictureDialogOpen,
+    isGroupPictureDialogOpen,
+    setIsGroupPictureDialogOpen,
+    newProfilePictureUrl,
+    setNewProfilePictureUrl,
+    newGroupPictureUrl,
+    setNewGroupPictureUrl,
+    profilePicture,
+    error,
+    setError,
+    socketError,
+    roomMembers,
+    userNames,
+    messageListRef,
+    handleSettingsClick,
+    handleSettingsClose,
+    handleDisconnect,
+    handleDeleteAccount,
+    handleSocketErrorClose,
+    handleUpdateProfilePicture,
+    handleUpdateGroupPicture,
+    handleLeaveGroup,
+    joinRoom,
+    sendMessage,
+    anchorEl,
+  } = useChatLogic();
 
-
-  useEffect(() => {
-    const fetchCurrentUserName = async () => {
-      if (userId) {
-        try {
-          const response = await API.get(`/users/${userId}`);
-          setCurrentUserName(response.data.username);
-          setUserNames(prev => ({...prev, [userId]: response.data.username}));
-        } catch (error) {
-          console.error("Error fetching current user's name:", error);
-        }
-      }
-    };
-  
-    fetchCurrentUserName();
-  }, [userId]);
-
-  useEffect(() => {
-    const newSocket = io(API_URL, {
-      withCredentials: true,
-      transports: ['websocket'],
-    });
-  
-    newSocket.on('connect', () => {
-      console.log('Connected to server');
-      setSocket(newSocket);
-    });
-  
-    newSocket.on('connect_error', (error) => {
-      console.error('Connection error:', error);
-      setSocketError('Failed to connect to the server');
-    });
-  
-    return () => {
-      newSocket.disconnect();
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!socket) return;
-  
-    const handleMessage = (message: Message) => {
-      setMessages((prevMessages) => [...prevMessages, message]);
-    };
-  
-    const handleMemberUpdate = (members: string[]) => {
-      setRoomMembers(members);
-    };
-  
-    socket.on('message', handleMessage);
-    socket.on('memberUpdate', handleMemberUpdate);
-  
-    return () => {
-      socket.off('message', handleMessage);
-      socket.off('memberUpdate', handleMemberUpdate);
-    };
-  }, [socket]);
-
-  useEffect(() => {
-    const storedUserId = localStorage.getItem('userId');
-    if (!storedUserId) {
-      navigate('/login');
-      return;
-    }
-    setUserId(storedUserId);
-    fetchUserProfilePicture(storedUserId);
-    fetchUserGroups(storedUserId);
-  }, [navigate]);
-
-  useEffect(() => { 
-    if (messageListRef.current) {
-      messageListRef.current.scrollTop = messageListRef.current.scrollHeight;
-    }
-  }, [messages]);
-
-  useEffect(() => {
-    const fetchMissingUserNames = async () => {
-      const missingUserIds = messages
-        .map(message => message.senderId)
-        .filter(senderId => senderId && senderId !== userId && !userNames[senderId]);
-  
-      const uniqueMissingUserIds = [...new Set(missingUserIds)];
-  
-      for (const senderId of uniqueMissingUserIds) {
-        await handleDifferentUserRender(senderId);
-      }
-    };
-  
-    fetchMissingUserNames();
-  }, [messages, userId, userNames]);
-
-  useEffect(() => {
-    const fetchMemberNames = async () => {
-      for (const memberId of roomMembers) {
-        if (memberId !== userId && !userNames[memberId]) {
-          await handleDifferentUserRender(memberId);
-        }
-      }
-    };
-  
-    fetchMemberNames();
-  }, [roomMembers]);
-
-  useEffect(() => {
-    console.log("userNames updated:", userNames);
-  }, [userNames]);
-
-
-  useEffect(() => {
-    console.log("userNames updated:", userNames);
-  }, [userNames]);
-  
-  useEffect(() => {
-    if (!socket) return;
-  
-    socket.on('memberLeft', ({ userId: leftUserId, groupId }) => {
-      setRoomMembers(prevMembers => prevMembers.filter(memberId => memberId !== leftUserId));
-      setGroups(prevGroups => prevGroups.map(group => 
-        group._id === groupId 
-          ? { ...group, members: group.members.filter(id => id !== leftUserId) } 
-          : group
-      ));
-  
-      if (leftUserId === userId) {
-        setSelectedRoom(null);
-        setMessages([]);
-      }
-    });
-  
-    socket.on('groupDeleted', ({ groupId }) => {
-      setGroups(prevGroups => prevGroups.filter(group => group._id !== groupId));
-      if (selectedRoom === groupId) {
-        setSelectedRoom(null);
-        setMessages([]);
-        setRoomMembers([]);
-      }
-    });
-  
-    return () => {
-      socket.off('memberLeft');
-      socket.off('groupDeleted');
-    };
-  }, [socket, selectedRoom, userId]);
-
-  const fetchUserGroups = async (userId: string) => {
-    try {
-      const groupIdsResponse = await API.get(`/users/${userId}/groups`);
-      const groupPromises = groupIdsResponse.data.map((groupId: string) => 
-        API.get(`/groups/${groupId}`)
-      );
-      const groupResponses = await Promise.all(groupPromises);
-      const fetchedGroups = groupResponses.map(response => response.data);
-      console.log('Fetched groups:', fetchedGroups);
-      setGroups(fetchedGroups);
-    } catch (error: any) {
-      console.error('Failed to fetch user groups:', error);
-      setError('Failed to fetch user groups');
-    }
-  };
-
-  const joinRoom = async (groupId: string) => {
-    if (!groupId) {
-      console.error('Invalid group ID');
-      setError('Invalid group selected');
-      return;
-    }
-    try {
-      socket?.emit('join', { userId, room: groupId });
-      socket?.once('joinSuccess', async ({ room, members }) => {
-        const messagesResponse = await API.get(`/messages/room/${room}`);
-        setMessages(messagesResponse.data);
-        setSelectedRoom(room);
-        setRoomMembers(members);
-  
-        for (const memberId of members) {
-          if (memberId !== userId && !userNames[memberId]) {
-            await handleDifferentUserRender(memberId);
-          }
-        }
-      });
-    } catch (error) {
-      console.log(error, 'Failed to join room');
-    }
-  };
-
-  const sendMessage = () => {
-    if (inputMessage.trim() === '' || !selectedRoom || !socket) return;
-    
-    const messageData = {
-      userId: userId,
-      room: selectedRoom,
-      content: inputMessage
-    };
-  
-    socket.emit('sendMessage', messageData);
-    setInputMessage('');
-  };
-
-  const handleSettingsClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleSettingsClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleDisconnect = () => {
-    handleSettingsClose();
-    socket?.disconnect();
-    localStorage.removeItem('userId');
-    localStorage.removeItem('profilePicture');
-    navigate('/login');
-  };
-
-  const handleDeleteAccount = async () => {
-    handleSettingsClose();
-    try {
-      await API.delete(`/users/${userId}`);
-      handleDisconnect();
-    } catch (error: any) {
-      setError('Failed to delete account');
-    }
-  };
-
-  const handleSocketErrorClose = (event?: React.SyntheticEvent | Event, reason?: string) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setSocketError(null);
-  };
-
-  const handleDifferentUserRender = async (senderId: string | undefined) => {
-    if (!senderId) {
-      console.log("Sender ID is undefined");
-      return "Unknown User";
-    }
-    
-    if (userNames[senderId]) {
-      return userNames[senderId];
-    }
-    
-    try {
-      const response = await API.get(`/users/${senderId}`);
-      const userName = response.data.username; 
-      setUserNames(prev => {
-        const newUserNames = {...prev, [senderId]: userName};
-        console.log("Updated userNames:", newUserNames);
-        return newUserNames;
-      });
-      return userName;
-    } catch (error) {
-      console.log("Error in handleDifferentUserRender", error);
-      return "Unknown User";
-    }
-  }
-  const handleUpdateProfilePicture = async () => {
-    try {
-      const response = await API.put(`/users/${userId}/profile-picture`, { profilePictureUrl: newProfilePictureUrl });
-      if (response.data && response.data.profilePicture) {
-        setProfilePicture(response.data.profilePicture);
-        localStorage.setItem('profilePicture', response.data.profilePicture);
-        setIsProfilePictureDialogOpen(false);
-        fetchUserProfilePicture(userId);
-      }
-    } catch (error) {
-      setError('Failed to update profile picture');
-    }
-  };
-  const fetchUserProfilePicture = async (userId: string) => {
-    try {
-      const response = await API.get(`users/${userId}/profile-picture`);
-      if (response.data && response.data.profilePicture) {
-        setProfilePicture(response.data.profilePicture);
-        localStorage.setItem('profilePicture', response.data.profilePicture);
-      }
-    } catch (error) {
-      console.log("error in fetchUserProfilePicture", error);
-    }
-  };
-
-  const handleUpdateGroupPicture = async () => {
-    if (!selectedRoom) return;
-    try {
-      const response = await API.put(`/groups/${selectedRoom}/group-picture`, { groupPictureUrl: newGroupPictureUrl });
-      if (response.data && response.data.groupPicture) {
-        setGroups(prevGroups => prevGroups.map(group => 
-          group._id === selectedRoom ? { ...group, groupPicture: response.data.groupPicture } : group
-        ));
-        setIsGroupPictureDialogOpen(false);
-      } else {
-        console.error('Group picture URL not found in the response');
-        setError('Failed to update group picture: URL not found in response');
-      }
-    } catch (error) {
-      console.error('Failed to update group picture:', error);
-      setError('Failed to update group picture');
-    }
-  };
-
-  const handleLeaveGroup = () => {
-    if (!selectedRoom || !socket) return;
-    socket.emit('leaveGroup', { userId, groupId: selectedRoom });
-    
-    setGroups(prevGroups => prevGroups.filter(group => group._id !== selectedRoom));
-    setSelectedRoom(null);
-    setMessages([]);
-    setRoomMembers([]);
-  };
-  
   return (
     <ChatContainer>
       <AppBar position="static">
